Authentication / Add a User Entity and fill it with dummy data
In general we follow the strategy outlined here:
http://symfony.com/doc/current/security/guard_authentication.html
Using Guard is a bit more difficult to setup but gives us more control.

It's important to note that this User class implements UserInterface
Introducing User as a Doctrine Entity user data as saved in the database.

User details could also be saved in a central authentication server.
In that case we'd still have a User class, but we would not store it in the DB with Doctrine.