Authentication / Render the login form
We'll deal with the Form processing later (separately).

The reason we create a \AppBundle\Form\LoginForm instead of creating a form in twig, as is customary for login forms (https://symfony.com/doc/current/security/form_login_setup.html), is to take advantage of the default usage of bootstrap for normal forms.

This is a rare case where we are not binding the loginForm with a class.