Authorization / Deny access for /admin
roles must start with ROLE_

There are two main ways to deny access:

- access_control (in security.yml - it's the simplest)
Symfony looks for a matching access control from top to bottom and stops as soon as it finds the first match.
(we can lock every page with an access_control and then whitelist the few publicpages with an access_control entry above that)
(we can also force the user to visit a specific part of the website via https)

- inside the controllers, using he authorization_checker service
a) We can use "createAccessDeniedException" if the user does not have the role we specify.
b) We can use "denyAccessUnlessGranted('ROLE_blabla')
c) We can use annotations (@Security("is_granted('ROLE_ADMIN')") either above a controller or above the whole class.