User Registration / Add conditional violation groups
By default, all constraints live within a group called "default"
When the form is validated, it validates all constraints in this "default" group.
When we use a different group, then when the form validates, it doesn't use that constraint.

We can specify which specific validation group the form will use when it tries to validate.
By doing so, we can make a constraint apply to one form but not another (when both are bound to the same entity)