Doctrine / Add unique column slug in genus
Since this column is unique, we need to follow one of the following two strategies:

A. make slug not unique at first, write a custom migration to generate unique slugs for all genuses and then make slug unique for new genuses. This approach makes a lot of sense when our DB is deployed on production.
B. make slug unique from the get go and recreate the whole DB. Drop the DB, create the DB again and run the migrations.