Doctrine / Extra fiels on the M-N relationship
Refactor the M-N to 2 M-1 relationships.
M-N relationships only work when we don't have any extra fields apart
from the 2 FKs on the Join table.
Otherwise we go for the explicit 2x M-1 relationships and an extra
Entity.

fetch="EXTRA_LAZY" works for any relationship that holds an array of
items, so we can keep it in genus.genusScientists

On the other hand we need to remove the JoinTable annotation on
genus.genusScientists as JoinTable and JoinColumn can only live on the
owning side of the relationship