Form / Readonly embedded field w/ form event listener
When a genusScientist is set in an embedded genus form, then we should
not be able to edit her. We should only be able to delete or add another
one. So, the genusScientist will become a read only in this case and
we'll display their email.

In Symfony language, this means that we want to remove the user field
from the embedded form.